+++ go.mod
@@ -1,0 +1,10 @@
+module Doodlering
+
+go 1.13
+
+require (
+	github.com/foolin/goview v0.3.0
+	github.com/gin-gonic/gin v1.6.3
+	github.com/go-sql-driver/mysql v1.5.0
+	github.com/gofrs/uuid v3.3.0+incompatible
+)
